src/market/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
requests
flask-marshmallow
cryptography
marshmallow-sqlalchemy
pymysql
